HSEARCH-4150 Move SimpleEntity to SearchQueryEntityChangingScrollingIT
Since no other test needs it.
